Move SymbolHider to tasks.internal
